Replace Octicons with Codicons
